requirements.txt
appdirs==1.4.4
astroid==2.4.2
black==20.8b1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.10
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==2.0.10
PyGithub==1.53
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
questionary==1.8.1
regex==2020.11.13
requests==2.25.0
rich==9.2.0
six==1.15.0
tabulate==0.8.7
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
wrapt==1.12.1
